Fixing XPath
DOGE updated their website that added an additional table that make the third div the fifth one and not the fourth one. This made the scraper unable to find the tables and thus broke. This should fix the scraper. This will git action this as its pushed to make sure it works right, I will commit later to comment this out again.