Debugging and Full Rework
So I accidently lied and said I would not be working on this, but here I am. This fixes an issue with xpaths changing on the DOGE site which broke the scraper and thus collection has not been done in a few weeks. This introduces a row by row collection of DOGE data. Previously the scraper collected data table page by table page. This method will take over 7 hours to complete and test but it gives every description of every contract, grant and lease.